One more de-dupe + move shorter/compact testing mock schemas to lib/data-mapper
