+++ libs/data-mapper/src/lib/__mocks__/SimpleMockSchema.json
