+++ libs/data-mapper/src/lib/__mocks__/index.ts
@@ -1,2 +1,3 @@
 export { default as simpleMockSchema } from './SimpleMockSchema.json';
 export { default as noChildrenMockSchema } from './NoChildrenMockSchema.json';
+export { default as heavyRepeatingMockSchema } from './HeavyRepeatingMockSchema.json';
