+++ libs/data-mapper/src/lib/components/breadcrumb/EditorBreadcrumb.stories.tsx
@@ -1,8 +1,8 @@
+import { simpleMockSchema } from '../../__mocks__';
 import { setCurrentTargetSchemaNode, setInitialSchema } from '../../core/state/DataMapSlice';
 import { store } from '../../core/state/Store';
 import type { Schema, SchemaExtended, SchemaNodeExtended } from '../../models/Schema';
 import { SchemaType } from '../../models/Schema';
-import { simpleMockSchema } from '../../models/__mocks__';
 import { convertSchemaToSchemaExtended } from '../../utils/Schema.Utils';
 import { EditorBreadcrumb } from './EditorBreadcrumb';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
