--- libs/data-mapper/src/lib/core/state/__mocks__/index.ts
@@ -1,1 +1,0 @@
-export { default as simpleMockSchema } from './SimpleMockSchema.json';
