+++ libs/data-mapper/src/lib/core/state/__test__/DataMapSlice.spec.ts
@@ -1,3 +1,4 @@
+import { simpleMockSchema } from '../../../__mocks__';
 import { concatFunction } from '../../../__mocks__/FunctionMock';
 import type { Schema, SchemaExtended } from '../../../models';
 import { SchemaType } from '../../../models';
@@ -17,7 +18,6 @@ import {
   deleteNodeFromConnections,
   deleteParentRepeatingConnections,
 } from '../DataMapSlice';
-import { simpleMockSchema } from '../__mocks__';
 import { sourceMockSchema } from '__mocks__/schemas';
 
 // NOTE: Functions with an unbounded input (like our concatFunction mock) will have two empty (undefined) values/fields by default
