+++ libs/data-mapper/src/lib/models/__test__/schema.spec.ts
@@ -1,6 +1,6 @@
+import { simpleMockSchema, noChildrenMockSchema } from '../../__mocks__';
 import { convertSchemaToSchemaExtended } from '../../utils/Schema.Utils';
 import type { Schema } from '../Schema';
-import { noChildrenMockSchema, simpleMockSchema } from '../__mocks__';
 
 describe('models/Schema', () => {
   describe('convertSchemaToSchemaExtended', () => {
