+++ libs/data-mapper/src/lib/utils/__mocks__/index.ts
@@ -1,4 +1,3 @@
-export { default as heavyRepeatingMockSchema } from './HeavyRepeatingMockSchema.json';
 export * from './linkedLoopConnections';
 export * from './indexedConnections';
 export * from './sourceSchemaNodes';
